Refactor Fabric to create Fabric component that can be reused
