+++ src/templates/csr/fabric/index.svelte
@@ -1,10 +1,7 @@
 <!-- https://polyfill.io/v3/polyfill.min.js?features=default -->
 <script lang="ts">
 	import type { GAMVariable } from '$lib/gam';
-	import { CLICK_MACRO, DEST_URL, isValidReplacedVariable } from '$lib/gam';
-	import { post } from '$lib/messenger';
-	import AdvertisementLabel from '$templates/components/AdvertisementLabel.svelte';
-	import Pixel from '$templates/components/Pixel.svelte';
+	import Fabric from '$templates/components/Fabric.svelte';
 
 	export let Trackingpixel: GAMVariable;
 	export let Researchpixel: GAMVariable;
@@ -33,124 +30,31 @@
 	export let MobileLayer2BackgroundPosition: GAMVariable;
 	export let MobileLayer3BackgroundImage: GAMVariable;
 	export let MobileLayer3BackgroundPosition: GAMVariable;
-
-	const isMobile = window.matchMedia('(max-width: 739px)').matches;
-	const isTablet = window.matchMedia(
-		'(min-width: 740px) and (max-width: 979px)',
-	).matches;
-
-	const [backgroundImage, backgroundPosition, backgroundRepeat] = isMobile
-		? [
-				MobileBackgroundImage,
-				MobileBackgroundImagePosition,
-				MobileBackgroundImageRepeat,
-			]
-		: [BackgroundImage, BackgroundImagePosition, BackgroundImageRepeat];
-
-	const checkScrollType = (scrollType: string): string =>
-		scrollType === 'parallax' && (isMobile || isTablet) ? 'fixed' : scrollType;
-
-	post({
-		type: 'background',
-		value: {
-			scrollType: checkScrollType(BackgroundScrollType),
-			backgroundColor: BackgroundColour,
-			backgroundImage: `url('${backgroundImage}')`,
-			backgroundRepeat,
-			backgroundPosition,
-		},
-	});
 </script>
 
-<svelte:body style:--background-color={BackgroundColour} />
-<AdvertisementLabel fullWidth={BackgroundScrollType === 'parallax'} />
-<a
-	class="fabric-container"
-	class:is-parallax={BackgroundScrollType === 'parallax'}
-	href={`${CLICK_MACRO}${DEST_URL}`}
-	target="_blank"
->
-	<div
-		class="layer"
-		style:--desktop-background-image={`url('${Layer1BackgroundImage}')`}
-		style:--desktop-background-position={Layer1BackgroundPosition}
-		style:--mobile-background-image={`url('${MobileLayer1BackgroundImage}')`}
-		style:--mobile-background-position={MobileLayer1BackgroundPosition}
-	/>
-	<div
-		class="layer"
-		style:--desktop-background-image={`url('${Layer2BackgroundImage}')`}
-		style:--desktop-background-position={Layer2BackgroundPosition}
-		style:--mobile-background-image={`url('${MobileLayer2BackgroundImage}')`}
-		style:--mobile-background-position={MobileLayer2BackgroundPosition}
-	/>
-	<div
-		class="layer"
-		style:--desktop-background-image={`url('${Layer3BackgroundImage}')`}
-		style:--desktop-background-position={Layer3BackgroundPosition}
-		style:--mobile-background-image={`url('${MobileLayer3BackgroundImage}')`}
-		style:--mobile-background-position={MobileLayer3BackgroundPosition}
-	/>
-</a>
-
-{#if isValidReplacedVariable(Trackingpixel)}
-	<Pixel src={Trackingpixel} />
-{:else if isValidReplacedVariable(Researchpixel)}
-	<Pixel src={Researchpixel} />
-{:else if isValidReplacedVariable(Viewabilitypixel)}
-	<Pixel src={Viewabilitypixel} />
-{/if}
-
-{@html thirdPartyJSTracking}
-
-<style lang="scss">
-	:global(body) {
-		background-color: var(--background-color);
-	}
-
-	.fabric-container {
-		display: block;
-		position: relative;
-		text-decoration: none;
-		margin: 0 auto;
-		height: 250px;
-		overflow: hidden;
-
-		&.is-parallax .layer {
-			background-size: auto;
-		}
-
-		@media (min-width: 740px) {
-			max-width: 740px;
-		}
-		@media (min-width: 980px) {
-			max-width: 980px;
-		}
-		@media (min-width: 1140px) {
-			max-width: 1140px;
-		}
-		@media (min-width: 1300px) {
-			max-width: 1300px;
-		}
-	}
-
-	.layer {
-		width: 100%;
-		height: 250px;
-		position: absolute;
-		top: 0;
-		left: 0;
-
-		background-repeat: no-repeat;
-		background-size: cover;
-		background-image: var(--mobile-background-image);
-		background-position: var(--mobile-background-position);
-	}
-
-	@media (min-width: 740px) {
-		.layer {
-			background-image: var(--desktop-background-image);
-			background-position: var(--desktop-background-position);
-		}
-	}
-</style>
+<Fabric
+	{Trackingpixel}
+	{Researchpixel}
+	{Viewabilitypixel}
+	{thirdPartyJSTracking}
+	{BackgroundScrollType}
+	{BackgroundColour}
+	{BackgroundImage}
+	{BackgroundImagePosition}
+	{BackgroundImageRepeat}
+	{MobileBackgroundImage}
+	{MobileBackgroundImagePosition}
+	{MobileBackgroundImageRepeat}
+	{Layer1BackgroundImage}
+	{Layer1BackgroundPosition}
+	{Layer2BackgroundImage}
+	{Layer2BackgroundPosition}
+	{Layer3BackgroundImage}
+	{Layer3BackgroundPosition}
+	{MobileLayer1BackgroundImage}
+	{MobileLayer1BackgroundPosition}
+	{MobileLayer2BackgroundImage}
+	{MobileLayer2BackgroundPosition}
+	{MobileLayer3BackgroundImage}
+	{MobileLayer3BackgroundPosition}
+/>
